refactor(parser): split pmc/parser/section.rs into per-element modules
`section.rs` was 1338 LOC and the #2 churn x complexity hotspot in the
workspace. Split it by JATS element group rather than by helper/driver
layer:

    section/mod.rs        entry point, SectionAction, SectionParts, scan_elements
    section/abstracts.rs  <abstract>
    section/body.rs       <body>, <sec>
    section/paragraph.rs  <p> and its inline children
    section/figure.rs     <fig>
    section/table.rs      <table-wrap>

Beyond the move, three decompositions carry the cognitive load that made
the file a hotspot:

- `extract_body_sections` mixed the sectioned-body loop with the
  "body without <sec>" fallback in one nested match. The tag dispatch
  moves to `classify_body_child` and the fallback accumulates into a
  `SectionParts`, dropping a nesting level.
- `read_paragraph_with_inline` handled inline elements inline. The
  `<xref>`/`<fig>`/`<table-wrap>` dispatch moves to `InlineContent`, so
  the paragraph loop only accumulates text and tracks depth. It now
  returns a named `Paragraph` instead of a 4-tuple.
- `SectionAction::ReadBodyParagraph` was byte-for-byte identical to
  `ReadParagraph`, and `EnterAbstract`/`SkipTitle` belonged to the
  abstract scanner, not the <sec> vocabulary. The abstract parser gets
  its own `AbstractAction`, `SectionAction` loses the dead variants, and
  `SectionParts::apply` no longer needs a "never arises" arm.

Also documents the intentional `SectionParts::apply` <->
`parse_section_from_body` recursion so cycle findings on the pair can be
dismissed on sight, and replaces the three copies of the fig/table
attribute lift with `FigAttrs::from_start` / `TableAttrs::from_start`.

Parser output is unchanged; all 26 section tests move with their
elements and pass unedited.

Closes #249
